Initial changes to introduce secretScanner class
